requirements.txt
python-dotenv==1.0.1
numpy==1.24.4
boto3==1.35.36
Flask==3.0.3
Flask_Cors==5.0.0
joblib==1.4.2
lightgbm==4.5.0
matplotlib==3.9.2
mlflow==2.17.0
dvc==3.53.0
dvc[s3]
nltk==3.9.1
pandas==2.2.3
wordcloud==1.9.3
seaborn==0.13.2
ipykernel
smote
optuna
xgboost

-e .